+++ patterns/dswp-vertical-card.php
@@ -1,0 +1,121 @@
+<?php
+/**
+ * Title: DSWP Vertical Card
+ * Slug: design-system-wordpress-theme/dswp-vertical-card
+ * Categories: columns
+ *
+ * @package Design-System-WordPress-Theme
+ */
+?>
+
+
+
+<!-- wp:group {"metadata":{"categories":["call-to-action"],"patternName":"design-system-wordpress-theme/card-with-hyperlinks-list","name":"Card with Hyperlink List"},"className":"alignfull","style":{"spacing":{"padding":{"top":"calc( 0.5 * var(\u002d\u002dwp\u002d\u002dstyle\u002d\u002droot\u002d\u002dpadding-right, var(\u002d\u002dwp\u002d\u002dcustom\u002d\u002dgap\u002d\u002dhorizontal)))","bottom":"calc( 0.5 * var(\u002d\u002dwp\u002d\u002dstyle\u002d\u002droot\u002d\u002dpadding-right, var(\u002d\u002dwp\u002d\u002dcustom\u002d\u002dgap\u002d\u002dhorizontal)))","left":"var(\u002d\u002dwp\u002d\u002dstyle\u002d\u002droot\u002d\u002dpadding-left, var(\u002d\u002dwp\u002d\u002dcustom\u002d\u002dgap\u002d\u002dhorizontal))","right":"var(\u002d\u002dwp\u002d\u002dstyle\u002d\u002droot\u002d\u002dpadding-right, var(\u002d\u002dwp\u002d\u002dcustom\u002d\u002dgap\u002d\u002dhorizontal))"},"margin":{"top":"0","bottom":"0"}}},"layout":{"type":"constrained"}} -->
+<div class="wp-block-group alignfull" style="margin-top:0;margin-bottom:0;padding-top:calc( 0.5 * var(--wp--style--root--padding-right, var(--wp--custom--gap--horizontal)));padding-right:var(--wp--style--root--padding-right, var(--wp--custom--gap--horizontal));padding-bottom:calc( 0.5 * var(--wp--style--root--padding-right, var(--wp--custom--gap--horizontal)));padding-left:var(--wp--style--root--padding-left, var(--wp--custom--gap--horizontal))"><!-- wp:group {"metadata":{"name":"Section"},"layout":{"type":"default"}} -->
+<div class="wp-block-group"><!-- wp:columns {"metadata":{},"align":"wide","className":"dswp-vertical-cards","style":{"spacing":{"blockGap":{"top":"var:preset|spacing|40","left":"0"},"padding":{"right":"0","left":"0"}}}} -->
+<div class="wp-block-columns alignwide dswp-vertical-cards" style="padding-right:0;padding-left:0"><!-- wp:column {"width":"","metadata":{"name":"Item"},"style":{"spacing":{"padding":{"top":"var:preset|spacing|10","bottom":"var:preset|spacing|10","left":"0","right":"0"}},"border":{"width":"0px","style":"none"}},"layout":{"type":"default"}} -->
+<div class="wp-block-column" style="border-style:none;border-width:0px;padding-top:var(--wp--preset--spacing--10);padding-right:0;padding-bottom:var(--wp--preset--spacing--10);padding-left:0"><!-- wp:group {"align":"full","className":"dswp-vertical-card","style":{"spacing":{"padding":{"right":"0","left":"0"}},"border":{"radius":"0px","color":"#d9d9d9","width":"1px"}},"layout":{"type":"constrained","contentSize":"111px","justifyContent":"left","wideSize":""}} -->
+<div class="wp-block-group alignfull dswp-vertical-card has-border-color" style="border-color:#d9d9d9;border-width:1px;border-radius:0px;padding-right:0;padding-left:0"><!-- wp:image {"id":15402,"sizeSlug":"full","linkDestination":"none","align":"full","className":"size-full"} -->
+<figure class="wp-block-image alignfull size-full"><img src="<?php echo esc_url( get_template_directory_uri() . '/assets/images/square-512.png' ); ?>" alt="" /></figure>
+<!-- /wp:image -->
+
+<!-- wp:group {"align":"full","layout":{"type":"constrained"}} -->
+<div class="wp-block-group alignfull"><!-- wp:heading {"textAlign":"left","level":5,"metadata":{"name":"Subtitle"},"className":"is-style-default","style":{"elements":{"link":{"color":{"text":"var:preset|color|typography-primary-color"}}},"spacing":{"padding":{"top":"0","bottom":"0"},"margin":{"top":"var:preset|spacing|50","bottom":"0"}}},"textColor":"typography-primary-color","fontSize":"xx-large"} -->
+<h5 class="wp-block-heading has-text-align-left is-style-default has-typography-primary-color-color has-text-color has-link-color has-xx-large-font-size" style="margin-top:var(--wp--preset--spacing--50);margin-bottom:0;padding-top:0;padding-bottom:0">Neque porro quisquam</h5>
+<!-- /wp:heading -->
+
+<!-- wp:paragraph {"style":{"spacing":{"margin":{"top":"var:preset|spacing|20","bottom":"var:preset|spacing|40","left":"0","right":"0"},"padding":{"right":"0","left":"0"}},"typography":{"lineHeight":"1.8"}}} -->
+<p style="margin-top:var(--wp--preset--spacing--20);margin-right:0;margin-bottom:var(--wp--preset--spacing--40);margin-left:0;padding-right:0;padding-left:0;line-height:1.8">Lorem ipsum dolor sit amet, consectetur adipiscing elit. Vivamus molestie a mi eu dictum. Sed eget dictum massa.</p>
+<!-- /wp:paragraph -->
+
+<!-- wp:paragraph {"style":{"spacing":{"margin":{"top":"0","bottom":"0"}}}} -->
+<p style="margin-top:0;margin-bottom:0"><a href="https://localhost/ticorp/test2/" data-type="page" data-id="15259">Neque porro quisquam</a></p>
+<!-- /wp:paragraph --></div>
+<!-- /wp:group -->
+
+<!-- wp:spacer {"height":"27px"} -->
+<div style="height:27px" aria-hidden="true" class="wp-block-spacer"></div>
+<!-- /wp:spacer --></div>
+<!-- /wp:group --></div>
+<!-- /wp:column -->
+
+<!-- wp:column {"width":"","metadata":{"name":"Item"},"style":{"spacing":{"padding":{"top":"var:preset|spacing|10","bottom":"var:preset|spacing|10","left":"0","right":"0"}},"border":{"width":"0px","style":"none"}},"layout":{"type":"default"}} -->
+<div class="wp-block-column" style="border-style:none;border-width:0px;padding-top:var(--wp--preset--spacing--10);padding-right:0;padding-bottom:var(--wp--preset--spacing--10);padding-left:0"><!-- wp:group {"align":"full","className":"dswp-vertical-card","style":{"spacing":{"padding":{"right":"0","left":"0"}},"border":{"radius":"0px","color":"#d9d9d9","width":"1px"}},"layout":{"type":"constrained","contentSize":"111px","justifyContent":"left","wideSize":""}} -->
+<div class="wp-block-group alignfull dswp-vertical-card has-border-color" style="border-color:#d9d9d9;border-width:1px;border-radius:0px;padding-right:0;padding-left:0"><!-- wp:image {"id":15402,"sizeSlug":"full","linkDestination":"none","align":"full","className":"size-full"} -->
+<figure class="wp-block-image alignfull size-full"><img src="<?php echo esc_url( get_template_directory_uri() . '/assets/images/square-512.png' ); ?>" alt="" /></figure>
+<!-- /wp:image -->
+
+<!-- wp:group {"align":"full","layout":{"type":"constrained"}} -->
+<div class="wp-block-group alignfull"><!-- wp:heading {"textAlign":"left","level":5,"metadata":{"name":"Subtitle"},"className":"is-style-default","style":{"elements":{"link":{"color":{"text":"var:preset|color|typography-primary-color"}}},"spacing":{"padding":{"top":"0","bottom":"0"},"margin":{"top":"var:preset|spacing|50","bottom":"0"}}},"textColor":"typography-primary-color","fontSize":"xx-large"} -->
+<h5 class="wp-block-heading has-text-align-left is-style-default has-typography-primary-color-color has-text-color has-link-color has-xx-large-font-size" style="margin-top:var(--wp--preset--spacing--50);margin-bottom:0;padding-top:0;padding-bottom:0">Neque porro quisquam</h5>
+<!-- /wp:heading -->
+
+<!-- wp:paragraph {"style":{"spacing":{"margin":{"top":"var:preset|spacing|20","bottom":"var:preset|spacing|40","left":"0","right":"0"},"padding":{"right":"0","left":"0"}},"typography":{"lineHeight":"1.8"}}} -->
+<p style="margin-top:var(--wp--preset--spacing--20);margin-right:0;margin-bottom:var(--wp--preset--spacing--40);margin-left:0;padding-right:0;padding-left:0;line-height:1.8">Lorem ipsum dolor sit amet, consectetur adipiscing elit. Vivamus molestie a mi eu dictum. Sed eget dictum massa.</p>
+<!-- /wp:paragraph -->
+
+<!-- wp:paragraph {"style":{"spacing":{"margin":{"top":"0","bottom":"0"}}}} -->
+<p style="margin-top:0;margin-bottom:0"><a href="https://localhost/ticorp/test2/" data-type="page" data-id="15259">Neque porro quisquam</a></p>
+<!-- /wp:paragraph --></div>
+<!-- /wp:group -->
+
+<!-- wp:spacer {"height":"27px"} -->
+<div style="height:27px" aria-hidden="true" class="wp-block-spacer"></div>
+<!-- /wp:spacer --></div>
+<!-- /wp:group --></div>
+<!-- /wp:column -->
+
+<!-- wp:column {"width":"","metadata":{"name":"Item"},"style":{"spacing":{"padding":{"top":"var:preset|spacing|10","bottom":"var:preset|spacing|10","left":"0","right":"0"}},"border":{"width":"0px","style":"none"}},"layout":{"type":"default"}} -->
+<div class="wp-block-column" style="border-style:none;border-width:0px;padding-top:var(--wp--preset--spacing--10);padding-right:0;padding-bottom:var(--wp--preset--spacing--10);padding-left:0"><!-- wp:group {"align":"full","className":"dswp-vertical-card","style":{"spacing":{"padding":{"right":"0","left":"0"}},"border":{"radius":"0px","color":"#d9d9d9","width":"1px"}},"layout":{"type":"constrained","contentSize":"111px","justifyContent":"left","wideSize":""}} -->
+<div class="wp-block-group alignfull dswp-vertical-card has-border-color" style="border-color:#d9d9d9;border-width:1px;border-radius:0px;padding-right:0;padding-left:0"><!-- wp:image {"id":15402,"sizeSlug":"full","linkDestination":"none","align":"full","className":"size-full"} -->
+<figure class="wp-block-image alignfull size-full"><img src="<?php echo esc_url( get_template_directory_uri() . '/assets/images/square-512.png' ); ?>" alt="" /></figure>
+<!-- /wp:image -->
+
+<!-- wp:group {"align":"full","layout":{"type":"constrained"}} -->
+<div class="wp-block-group alignfull"><!-- wp:heading {"textAlign":"left","level":5,"metadata":{"name":"Subtitle"},"className":"is-style-default","style":{"elements":{"link":{"color":{"text":"var:preset|color|typography-primary-color"}}},"spacing":{"padding":{"top":"0","bottom":"0"},"margin":{"top":"var:preset|spacing|50","bottom":"0"}}},"textColor":"typography-primary-color","fontSize":"xx-large"} -->
+<h5 class="wp-block-heading has-text-align-left is-style-default has-typography-primary-color-color has-text-color has-link-color has-xx-large-font-size" style="margin-top:var(--wp--preset--spacing--50);margin-bottom:0;padding-top:0;padding-bottom:0">Neque porro quisquam</h5>
+<!-- /wp:heading -->
+
+<!-- wp:paragraph {"style":{"spacing":{"margin":{"top":"var:preset|spacing|20","bottom":"var:preset|spacing|40","left":"0","right":"0"},"padding":{"right":"0","left":"0"}},"typography":{"lineHeight":"1.8"}}} -->
+<p style="margin-top:var(--wp--preset--spacing--20);margin-right:0;margin-bottom:var(--wp--preset--spacing--40);margin-left:0;padding-right:0;padding-left:0;line-height:1.8">Lorem ipsum dolor sit amet, consectetur adipiscing elit. Vivamus molestie a mi eu dictum. Sed eget dictum massa.</p>
+<!-- /wp:paragraph -->
+
+<!-- wp:paragraph {"style":{"spacing":{"margin":{"top":"0","bottom":"0"}}}} -->
+<p style="margin-top:0;margin-bottom:0"><a href="https://localhost/ticorp/test2/" data-type="page" data-id="15259">Neque porro quisquam</a></p>
+<!-- /wp:paragraph --></div>
+<!-- /wp:group -->
+
+<!-- wp:spacer {"height":"27px"} -->
+<div style="height:27px" aria-hidden="true" class="wp-block-spacer"></div>
+<!-- /wp:spacer --></div>
+<!-- /wp:group --></div>
+<!-- /wp:column -->
+
+<!-- wp:column {"width":"","metadata":{"name":"Item"},"style":{"spacing":{"padding":{"top":"var:preset|spacing|10","bottom":"var:preset|spacing|10","left":"0","right":"0"}},"border":{"width":"0px","style":"none"}},"layout":{"type":"default"}} -->
+<div class="wp-block-column" style="border-style:none;border-width:0px;padding-top:var(--wp--preset--spacing--10);padding-right:0;padding-bottom:var(--wp--preset--spacing--10);padding-left:0"><!-- wp:group {"align":"full","className":"dswp-vertical-card","style":{"spacing":{"padding":{"right":"0","left":"0"}},"border":{"radius":"0px","color":"#d9d9d9","width":"1px"}},"layout":{"type":"constrained","contentSize":"111px","justifyContent":"left","wideSize":""}} -->
+<div class="wp-block-group alignfull dswp-vertical-card has-border-color" style="border-color:#d9d9d9;border-width:1px;border-radius:0px;padding-right:0;padding-left:0"><!-- wp:image {"id":15402,"sizeSlug":"full","linkDestination":"none","align":"full","className":"size-full"} -->
+<figure class="wp-block-image alignfull size-full"><img src="<?php echo esc_url( get_template_directory_uri() . '/assets/images/square-512.png' ); ?>" alt="" /></figure>
+<!-- /wp:image -->
+
+<!-- wp:group {"align":"full","layout":{"type":"constrained"}} -->
+<div class="wp-block-group alignfull"><!-- wp:heading {"textAlign":"left","level":5,"metadata":{"name":"Subtitle"},"className":"is-style-default","style":{"elements":{"link":{"color":{"text":"var:preset|color|typography-primary-color"}}},"spacing":{"padding":{"top":"0","bottom":"0"},"margin":{"top":"var:preset|spacing|50","bottom":"0"}}},"textColor":"typography-primary-color","fontSize":"xx-large"} -->
+<h5 class="wp-block-heading has-text-align-left is-style-default has-typography-primary-color-color has-text-color has-link-color has-xx-large-font-size" style="margin-top:var(--wp--preset--spacing--50);margin-bottom:0;padding-top:0;padding-bottom:0">Neque porro quisquam</h5>
+<!-- /wp:heading -->
+
+<!-- wp:paragraph {"style":{"spacing":{"margin":{"top":"var:preset|spacing|20","bottom":"var:preset|spacing|40","left":"0","right":"0"},"padding":{"right":"0","left":"0"}},"typography":{"lineHeight":"1.8"}}} -->
+<p style="margin-top:var(--wp--preset--spacing--20);margin-right:0;margin-bottom:var(--wp--preset--spacing--40);margin-left:0;padding-right:0;padding-left:0;line-height:1.8">Lorem ipsum dolor sit amet, consectetur adipiscing elit. Vivamus molestie a mi eu dictum. Sed eget dictum massa.</p>
+<!-- /wp:paragraph -->
+
+<!-- wp:paragraph {"style":{"spacing":{"margin":{"top":"0","bottom":"0"}}}} -->
+<p style="margin-top:0;margin-bottom:0"><a href="https://localhost/ticorp/test2/" data-type="page" data-id="15259">Neque porro quisquam</a></p>
+<!-- /wp:paragraph --></div>
+<!-- /wp:group -->
+
+<!-- wp:spacer {"height":"27px"} -->
+<div style="height:27px" aria-hidden="true" class="wp-block-spacer"></div>
+<!-- /wp:spacer --></div>
+<!-- /wp:group --></div>
+<!-- /wp:column --></div>
+<!-- /wp:columns --></div>
+<!-- /wp:group --></div>
+<!-- /wp:group -->